+++ package.json
@@ -9,6 +9,7 @@
     "@elastic/search-ui-app-search-connector": "^1.10.1",
     "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.6.0",
     "@mui/material": "^5.5.0",
     "body-scroll-lock": "^3.1.5",
     "react": "^17.0.2",
