+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Header from "./components/header";
+import Header from "./components/Header";
 import "./App.css";
 
 import AppSearchAPIConnector from "@elastic/search-ui-app-search-connector";
